requirements/base.txt
asgiref==3.3.1
Django==3.1.7
django-cors-headers==3.7.0
django-filter==2.4.0
djangorestframework==3.12.2
python-decouple==3.4
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
